separated watchlist dialog modal functionality from shared dialog modal file
