+++ js/render/home-page/movie-list-page.js
@@ -9,10 +9,10 @@ import {
     clearDialogModal, 
     setDialogModalMessage, 
     setDialogModalType, 
-    getConfirmationDialogModalStatusFromLocalStorage, 
-    setConfirmationDialogModalStatusInLocalStorage,
-    initWatchlistDialogModals
+    getConfirmationModalStatus, 
+    setConfirmationModalStatus,
 } from "../shared/dialog-modal.js";
+import { initWatchlistDialogModals } from "../shared/watchlist-dialog-modal.js";
 import { renderSearchForm } from "../shared/form.js";
 import { initMovieInfoModal } from "../shared/movie-info.modal.js";
 
@@ -20,7 +20,7 @@ async function initMovieListPage() {
     const $container = document.querySelector("#container");
     const urlSearchParams = new URLSearchParams(window.location.search);
     const searchResult = urlSearchParams.get("search");
-    const shouldShowConfirmationDialog = getConfirmationDialogModalStatusFromLocalStorage();
+    const shouldShowConfirmationDialog = getConfirmationModalStatus();
 
     renderLoading();
     renderMenuNav();
@@ -38,7 +38,7 @@ async function initMovieListPage() {
     if (shouldShowConfirmationDialog) {
         const confirmationMessage = setDialogModalMessage("You have been successfully signed out");
         const modalType = setDialogModalType("confirmation-modal");
-        setConfirmationDialogModalStatusInLocalStorage(false);
+        setConfirmationModalStatus(false);
         intentionalDelay(() => {
             clearLoading();
             renderDialogModal(confirmationMessage, modalType);
